Refactoring: GeoDataRiverSurvey::InsertRiverPathPointCommand is moved to separated files.
