+++ libs/geodata/riversurvey/riversurvey.pro
@@ -120,6 +120,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_changeselectioncommand.h \
            private/geodatariversurvey_deleteriverpathpointcommand.h \
            private/geodatariversurvey_expandcrosssectioncommand.h \
+           private/geodatariversurvey_insertriverpathpointcommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
            private/geodatariversurveyproxy_setsettingcommand.h
@@ -166,6 +167,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_changeselectioncommand.cpp \
            private/geodatariversurvey_deleteriverpathpointcommand.cpp \
            private/geodatariversurvey_expandcrosssectioncommand.cpp \
+           private/geodatariversurvey_insertriverpathpointcommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveyproxy_setsettingcommand.cpp
 RESOURCES += riversurvey.qrc
